proxy and cache download from gihub during test
