+++ test/Certes.Tests.Web/Startup.cs
@@ -1,6 +1,8 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.IO;
+using System.Linq;
+using System.Net.Http;
 using System.Text;
 using System.Threading.Tasks;
 using Certes.Jws;
@@ -15,11 +17,15 @@ using Microsoft.Azure.Management.ResourceManager.Fluent.Authentication;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.Logging;
 using Newtonsoft.Json;
+using Newtonsoft.Json.Linq;
 
 namespace Certes.Tests.Web
 {
     public class Startup
     {
+        private string CertificateData { get; set; }
+        private string CertificateKey { get; set; }
+
         public IConfigurationRoot Configuration { get; }
 
         public Startup(IHostingEnvironment env)
@@ -34,7 +40,7 @@ namespace Certes.Tests.Web
             builder.AddEnvironmentVariables();
             Configuration = builder.Build();
         }
-        
+
         public void Configure(IApplicationBuilder app, IHostingEnvironment env, ILoggerFactory loggerFactory)
         {
             loggerFactory.AddConsole();
@@ -46,7 +52,57 @@ namespace Certes.Tests.Web
 
             app.Map("/dns-01", sub => sub.Run(Dns));
             app.Map("/.well-known/acme-challenge", sub => sub.Run(AcmeChallenge));
+            app.Map("/cert-data", sub => sub.Run(CertData));
+            app.Map("/cert-key", sub => sub.Run(CertKey));
             app.Run(Fallback);
+        }
+
+        private async Task CertData(HttpContext context)
+        {
+            if (CertificateData == null)
+            {
+                using (var http = new HttpClient())
+                {
+                    http.DefaultRequestHeaders.UserAgent.ParseAdd("request");
+                    var json = await http.GetStringAsync("https://api.github.com/repos/fszlin/lo0.in/releases/latest");
+                    var metadata = JObject.Parse(json);
+
+                    var certUrl = metadata["assets"]
+                        .AsJEnumerable()
+                        .Where(a => a["name"].Value<string>() == "cert.pem")
+                        .Select(a => a["browser_download_url"].Value<string>())
+                        .First();
+
+                    CertificateData = await http.GetStringAsync(certUrl);
+                }
+            }
+
+            context.Response.ContentType = "application/x-pem-file";
+            await context.Response.WriteAsync(CertificateData);
+        }
+
+        private async Task CertKey(HttpContext context)
+        {
+            if (CertificateKey == null)
+            {
+                using (var http = new HttpClient())
+                {
+                    http.DefaultRequestHeaders.UserAgent.ParseAdd("request");
+                    var json = await http.GetStringAsync("https://api.github.com/repos/fszlin/lo0.in/releases/latest");
+                    var metadata = JObject.Parse(json);
+
+                    var keyUrl = metadata["assets"]
+                        .AsJEnumerable()
+                        .Where(a => a["name"].Value<string>() == "key.pem")
+                        .Select(a => a["browser_download_url"].Value<string>())
+                        .First();
+
+                    CertificateKey = await http.GetStringAsync(keyUrl);
+                }
+            }
+
+            context.Response.ContentType = "application/x-pem-file";
+            await context.Response.WriteAsync(CertificateKey);
         }
 
         private async Task Dns(HttpContext context)
