+++ test/Certes.Tests/Helper.cs
@@ -98,37 +98,13 @@ namespace Certes
                 return validCertificate.Value;
             }
 
-            for (int i = 0; i < 10; ++i)
+            using (var http = new HttpClient())
             {
-                try
-                {
-                    using (var http = new HttpClient())
-                    {
-                        http.DefaultRequestHeaders.UserAgent.ParseAdd("request");
-                        var json = await http.GetStringAsync("https://api.github.com/repos/fszlin/lo0.in/releases/latest");
-                        var metadata = JObject.Parse(json);
+                var cert = await http.GetStringAsync("http://certes-ci.dymetis.com/cert-data");
+                var key = await http.GetStringAsync("http://certes-ci.dymetis.com/cert-key");
 
-                        var certUrl = metadata["assets"]
-                            .AsJEnumerable()
-                            .Where(a => a["name"].Value<string>() == "cert.pem")
-                            .Select(a => a["browser_download_url"].Value<string>())
-                            .First();
-                        var keyUrl = metadata["assets"]
-                            .AsJEnumerable()
-                            .Where(a => a["name"].Value<string>() == "key.pem")
-                            .Select(a => a["browser_download_url"].Value<string>())
-                            .First();
-
-                        var cert = await http.GetStringAsync(certUrl);
-                        var key = await http.GetStringAsync(keyUrl);
-
-                        validCertificate = (cert, key);
-                        return validCertificate.Value;
-                    }
-                }
-                catch
-                {
-                }
+                validCertificate = (cert, key);
+                return validCertificate.Value;
             }
 
             throw new Exception("Fail to load certificate");
